# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ServiceResponse_any-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ServiceResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CompletedActivity": {
				"properties": {
					"project": {
						"type": "string"
					},
					"task": {
						"type": "string"
					}
				},
				"required": [
					"project",
					"task"
				],
				"type": "object",
				"additionalProperties": false
			},
			"NextWeekTask": {
				"properties": {
					"task": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					}
				},
				"required": [
					"task",
					"startDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WeeklyReportFormData": {
				"properties": {
					"supervisorName": {
						"type": "string"
					},
					"staffName": {
						"type": "string"
					},
					"periodFrom": {
						"type": "string"
					},
					"periodTo": {
						"type": "string"
					},
					"hoursWorked": {
						"type": "string"
					},
					"completedActivities": {
						"items": {
							"$ref": "#/components/schemas/CompletedActivity"
						},
						"type": "array"
					},
					"challenges": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"nextWeekTasks": {
						"items": {
							"$ref": "#/components/schemas/NextWeekTask"
						},
						"type": "array"
					},
					"immediateIssues": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"supervisorName",
					"staffName",
					"periodFrom",
					"periodTo",
					"hoursWorked",
					"completedActivities",
					"challenges",
					"nextWeekTasks",
					"immediateIssues"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_WeeklyReportFormData_": {
				"properties": {
					"supervisorName": {
						"type": "string"
					},
					"staffName": {
						"type": "string"
					},
					"periodFrom": {
						"type": "string"
					},
					"periodTo": {
						"type": "string"
					},
					"hoursWorked": {
						"type": "string"
					},
					"completedActivities": {
						"items": {
							"$ref": "#/components/schemas/CompletedActivity"
						},
						"type": "array"
					},
					"challenges": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"nextWeekTasks": {
						"items": {
							"$ref": "#/components/schemas/NextWeekTask"
						},
						"type": "array"
					},
					"immediateIssues": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"optionalJwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "community-backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Ikundabayo placide"
		}
	},
	"paths": {
...
		"/api/weeklyreports": {
			"get": {
				"operationId": "ListWeeklyReports",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"WeeklyReports"
				],
				"security": [
					{
						"jwt": [
							"weeklyreport:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 25,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"pending",
								"reviewed",
								"approved",
								"rejected",
								"seen"
							]
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "startDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "endDate",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateWeeklyReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"WeeklyReports"
				],
				"security": [
					{
						"jwt": [
							"weeklyreport:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/WeeklyReportFormData"
							}
						}
					}
				}
			}
		},
		"/api/weeklyreports/{id}": {
			"get": {
				"operationId": "GetWeeklyReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"WeeklyReports"
				],
				"security": [
					{
						"jwt": [
							"weeklyreport:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateWeeklyReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"WeeklyReports"
				],
				"security": [
					{
						"jwt": [
							"weeklyreport:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/Partial_WeeklyReportFormData_"
									},
									{
										"properties": {
											"reviewComments": {
												"type": "string"
											},
											"status": {
												"type": "string",
												"enum": [
													"pending",
													"reviewed",
													"approved",
													"rejected",
													"seen"
												]
											}
										},
										"type": "object"
									}
								]
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteWeeklyReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"WeeklyReports"
				],
				"security": [
					{
						"jwt": [
							"weeklyreport:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/weeklyreports/user/{userId}": {
			"get": {
				"operationId": "GetUserWeeklyReports",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"WeeklyReports"
				],
				"security": [
					{
						"jwt": [
							"weeklyreport:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 25,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"pending",
								"reviewed",
								"approved",
								"rejected",
								"seen"
							]
						}
					}
				]
			}
		},
		"/api/weeklyreports/{id}/review": {
			"put": {
				"operationId": "MarkAsReviewed",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"WeeklyReports"
				],
				"security": [
					{
						"jwt": [
							"weeklyreport:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/weeklyreports/{id}/seen": {
			"put": {
				"operationId": "MarkAsSeen",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"WeeklyReports"
				],
				"security": [
					{
						"jwt": [
							"weeklyreport:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}